+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [2.0.0]
 ### Added
 - Introduce a global exposure limit to restrict logging output across libraries. The limit
   defaults to `.critical` and must be configured at startup to enable more verbose logging.